controller/requirements.txt
# FastAPI and ASGI server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Data validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Configuration parsing
pyyaml>=6.0
toml>=0.10.2

# Process management
psutil>=5.9.0

# Prometheus metrics
prometheus-client>=0.19.0

# WebSocket support
websockets>=12.0

# Template engine (for Liquidsoap scripts)
jinja2>=3.1.0

# Async file operations
aiofiles>=23.0.0

# File uploads
python-multipart>=0.0.6

# HTTP client
httpx>=0.25.0

# Logging
python-json-logger>=2.0.7
